Port to CouchDB: Rename src/riaknostic*
Part 2/8

This commit renames the riaknostic files which are to be ported
to weatherreport and adds a modification note to the notices at
the top of each file.

This is groundwork for the following commits where the
functionality of each file will be modified to work with Apache
CouchDB.

It should be noted that this commit completely breaks the build.