Generalise threshold checks on process attributes
This commit generalises the logic that checks for process
attributes and moves it to weatherreport_utils. This is
groundwork for the following commit which will add a check
for processes by memory usage by re-using the generalised code.

We also take the opportunity to rename fold_processes to
something that is more descriptive and also correct.

BugzID: 32875