Remove redundant rpc calls
The initial design of Riaknostic (and hence WeatherReport) was
that checks ran in the script and made RPC calls to the local
cluster node when needed. Due to a requirement to allow checks
to run on multiple cluster nodes without guarantees of SSH
connectivity, the design has changed such that each check is
now run entirely on the cluster node.

This means it is no longer necessary to use
weatherreport_node:local_command/3 to interact with the cluster,
the calls can just be made directly.

This commit removes the redundant calls to
weatherreport_node:local_command/3 which makes the intent of the
code a bit clearer.

BugzID: 34016